requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
trimesh>=3.9.0
scipy>=1.7.0
matplotlib>=3.4.0
pillow>=8.3.0
pyglet<2.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-Login==0.6.3
Pillow==10.0.0
reportlab==4.0.4
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
bcrypt==4.0.1
email-validator==2.1.0
